Extract message routing
